refactor: split off the tui code
